+++ crypto/elliptic-curve-endomorphism.ts
